requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
pytz
timezonefinder
google-generativeai
